requirements.txt
ccxt
requests
pandas
matplotlib
mplfinance
python-dotenv